+++ SpatialGDK/Source/SpatialGDKEditor/Public/SpatialGDKEditorSettings.h
@@ -328,6 +328,9 @@ public:
 	UPROPERTY(EditAnywhere, config, Category = "Launch", meta = (DisplayName = "Exposed local runtime IP address"))
 	FString ExposedRuntimeIP;
 
+	UPROPERTY(EditAnywhere, config, Category = "Launch", meta = (DisplayName = "Stop local deployment on stop play in editor"))
+	bool bStopLocalDeploymentOnEndPIE;
+
 	/** Select the check box to stop your game’s local deployment when you shut down Unreal Editor. */
 	UPROPERTY(EditAnywhere, config, Category = "Launch", meta = (DisplayName = "Stop local deployment on exit"))
 	bool bStopSpatialOnExit;
@@ -423,10 +426,6 @@ public:
 	/** The token created using 'spatial project auth dev-auth-token' */
 	UPROPERTY(EditAnywhere, config, Category = "Cloud Connection")
 	FString DevelopmentAuthenticationToken;
-
-	/** The deployment to connect to when using the Development Authentication Flow. If left empty, it uses the first available one (order not guaranteed when there are multiple items). The deployment needs to be tagged with 'dev_login'. */
-	UPROPERTY(EditAnywhere, config, Category = "Cloud Connection")
-	FString DevelopmentDeploymentToConnect;
 
 private:
 	UPROPERTY(config)
@@ -576,7 +575,7 @@ public:
 	{
 		if (!IsRegionCodeValid(PrimaryDeploymentRegionCode))
 		{
-			return FText::FromString(TEXT("Invalid"));
+			return NSLOCTEXT("SpatialGDKEditorSettings", "InvalidRegion", "Invalid");
 		}
 
 		UEnum* Region = FindObject<UEnum>(ANY_PACKAGE, TEXT("ERegionCode"), true);
@@ -612,7 +611,7 @@ public:
 	{
 		if (!IsRegionCodeValid(SimulatedPlayerDeploymentRegionCode))
 		{
-			return FText::FromString(TEXT("Invalid"));
+			return NSLOCTEXT("SpatialGDKEditorSettings", "InvalidRegion", "Invalid");
 		}
 
 		UEnum* Region = FindObject<UEnum>(ANY_PACKAGE, TEXT("ERegionCode"), true);
@@ -697,8 +696,8 @@ public:
 	bool IsDeploymentConfigurationValid() const;
 
 	void SetDevelopmentAuthenticationToken(const FString& Token);
-	void SetDevelopmentDeploymentToConnect(const FString& Deployment);
 
+	static bool IsValidIP(const FString& IP);
 	void SetExposedRuntimeIP(const FString& RuntimeIP);
 
 	void SetSpatialOSNetFlowType(ESpatialOSNetFlow::Type NetFlowType);